Update patch set 1
Patch Set 1: Code-Review+2

Thanks Matt on your contribution.
This code is inline with the existing ovsdb code-base.
But ofcourse the existing infra has its kinks and we are ironing it out in the netty branch.

So, though we are pushing this in, it will take some massaging around this functionality in the netty branch.
Hopefully it will be much more streamlined and nice to look at @ :)

Patch-set: 1
Label: Code-Review=+2
Label: SUBM=+1