Update patch set 1
Patch Set 1:

Totally understood, Madhu. For me this was mostly about getting familiar with 1) the way things are done on ODL, 2) the structure of ODL and 3) how OVSDB works. So this definitely moved me in that direction.

Thanks to both of you for the help on this one and for the opportunity to learn.

Patch-set: 1
Reviewer: Gerrit User 897 <897@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0