Update patch set 1
Patch Set 1: Code-Review+2

This looks good for the functionality it provides.
This will be modified 2 ways :
1. Netty based
2. the Controller configuration will be automatic during the bridge creation & new controller addition to the cluster there off.

Due to the work in the netty branch, we have decided to bring in all such functionality related changes to the master so that we can  pull them in during the collapse.

Approved.

Patch-set: 1
Label: Code-Review=+2