Added external_id bridgeDom support for ovsdb
Applied patch from Madhu. Thanks.

Added a column entries to get review
for needing a generic parameter for
set <table> <column> <param>.
Junit will update the external_id column in
the Bridge table. The method will be modified
to accept columns as opposed to
"external_id" only. Netty branch addresses
this with POJOs but this is for current
functionality.

Signed-off-by: Brent Salisbury <brent.salisbury@gmail.com>
Change-Id: Iba1913a91dea258541d16ca348179bc541a7422f