Update patch set 1
Patch Set 1:

(1 comment)

In addition to the inline comment, as a generic comment, it ia always a good idea to amend the existing gerrit with the changes done on top of it. This will help avoid lots of merge issues and help in tracking.

So, Can you please abandon this and amend the previous commit ?

Patch-set: 1
Label: Verified=0