Minor fix to make Junit happy.
Due to the introduction of the passive connection support, the Monitor and Schema was pushed to a
separate thread. Also the connection was deemed successful only if the monitor was successful.
That doesnt work well with the current (buggy) implemenation of JUnit. But since the JUnit is very
handy in testing out the functionality, reversing the ovsdbConnection management.

Change-Id: Ifa833014c587e94baa63a5925959c87d9e27c6fc
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>