Controller IP-Address filtered using the "of.address" controller configuration from a list of all the exposed ip-addresses of the system.
As per the current implementation, when a new bridge is added, the Configuration Service automatically sets up the OF Controller field
of the bridge to point back at the OpenDaylight Openflow plugin. This is done by scanning all the open interfaces with valid IP-Addresses
and sets it in the Bridge table's controller column.
This can be better handled by filtering it further by using just the Admin prefered ip-address of the controller.
This configuration item is called "of.address" in configuration/config.ini.

Change-Id: Icdfeb0dfc4e68e61e926f09b49461fe0c1dd3060
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>