Disabling the installation of LLDP Punt rules to Controller.
In a multi-node devstack setup, LLDP installation seems to cause issues with the Controller
(CPU spike and some OF message timeouts). It is better to debug that and once we have an idea,
we can enable it again. There is NO functional impact to the Neutron integration with OVSDB
by disabling LLDP flow installation.

Change-Id: I51e64253e09b6e863e2eea89f2570e436d3cc93f
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>