Update patch set 3
Patch Set 3:

This is good change.
But the Status with Description as UUID is used beyond the OVSDBNorthbound.java. It is currently used heavily in ovsdb.neutron bundle as well.

Java files in ovsdb.neutron bundle must be changed to accommodate this else, it will break neutron functionalities.

Patch-set: 3
Label: Verified=0