Made changes in Internal Network creation to also takes care of creating br-int & its patch-ports if they are missing.
Please note that the current implementation creates br-int, br-tun and the patch ports upon detecting a new Node.
This can be improved by creating these bridges just when the Neutron events begin. Since this functionality is now
boxed into a single method call, it can be taken care as a future fix.

Change-Id: I71f52348d042a8ef1653fcd8e094a7a1d87d78d6
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>