Installing NORMAL rules on all the default neutron bridges.
Opendaylight controller's OF1.0 plugin deletes all the OF rules upon initial connection.
This could be a valid scenario on regular scenario, but it breaks some basic assumptions
in a Openstack/devstack deployment. Hence this fix will override that behaviour by installing
a NORMAL rule as the default rule to avoid data traffic to be punted to the controller.
This is a temporary measure to solves some major issues seen on couple of deployment scenarios.
We will address it more elagantly once we integrate with the new OF plugin.

Change-Id: Ie53c85c3e1549d07392d6c8129adb05806ed32cd
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>