Moved all Openflow 1.0 specific flow programming away from InternalNetworkManager to OF10ProviderManager class.
Also, handled Switchmanager Node coming up event to take care of initializing the default flows.
This is needed to handle timing issues in which the ForwardingRulesManager rejects request to program
default flows from ovsdb.neutron if the request is too soon (before the OF node gets registered with other
OSGi bundles).

Change-Id: Id299c75146e504df63a45b77cc0f6be962c3abb6
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>