Aligning Jackson to the Controller's l latest upgrade (2.3.0). Also updated the northbound bundle to use the latest Jackson.
Due to this change, the ovsdb.northbound bundle will start to accept and respond OVSDB JSON messages in the same format as
expected by the ovsdb-server. (i.e) a set will be represented as ["set",[]] instead of the common sense [].
This is due to the ovsdb requirement as specified in the RFC / Draft.

Change-Id: I3a1a2548f1b8c87d1d4b63f5952d6fabaa6f3d33
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>