Make automagically adding a controller to each detected bridge after connecting an OVS node a configurable setting. The default is ovsdb.autoconfigurecontroller=true
I've seen multiple configurations where the ip address of a host is configured on the bridge interface.
So for example there is one bridge (cloudbr0) and one interface (eth0). Eth0 would be added as a port to cloudbr0 and the ip address of the host would be configured on cloudbr0. This is actually the recommended configuration when using CloudStack with a Xen or KVM hypervisor running OVS. I've seen this configuration is multiple other setups as well.
When a node with this configuration is added to the ODL controller, all communication will be blocked as a controller implies that the switch will be flow based. This causes ovsdb to lose communication to the outside world and effectively isolates the hypervisor from the network.

This commit adds a property to configure this feature from config.ini. Set "ovsdb.autoconfigurecontroller" to true to configure controllers on all detected openvswitch bridges.

Change-Id: I8285bd2b8a7000997b4b62373bd9e02457c93d8e
Signed-off-by: Hugo Trippaers <hugo@trippaers.nl>