Update patch set 1
Patch Set 1:

This might require some more explanation. I've seen multiple configurations where the ip address of a host is configured on the bridge interface.

So for example there is one bridge (cloudbr0) and one interface (eth0). Eth0 would be added as a port to cloudbr0 and the ip address of the host would be configured on cloudbr0. This is actually the recommended configuration when using CloudStack with a Xen or KVM hypervisor running OVS. I've seen this configuration is multiple other setups as well.

When a node with this configuration is added to the ODL controller, all communication will be blocked as a controller implies that the switch will be flow based. This causes ovsdb to lose communication to the outside world and effectively isolates the hypervisor from the network.

Patch-set: 1
Label: Verified=0