Update patch set 1
Patch Set 1:

Hugo, after the patch is merged, the comment with the explanation will be left in gerrit, but developers pulling it and doing git log will not see it.
It's better to add it as part of the commit message that will be preserved in git.

Patch-set: 1
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0