Update patch set 1
Patch Set 1:

Based on the discussion we had on the release meeting, We decided not to pursue this Bug if it only applicable to the Test cases.

Since the changes are limited to UT and IT. I would rather not merge this.

If you can spot anything on the nonTest files, please submit them and we can merge it.

Patch-set: 1
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0