Update patch set 2
Patch Set 2: Code-Review+2

Thanks Brent for pushing the code.
We have some work to be done in making sure it fits the
integration. But, at this point, i am inclined to merge
this in as our manual testing results are encouraging.

We shall take up the bug fixes and other integration pieces
in smaller commits that will follow.

Patch-set: 2
Label: Code-Review=+2
Label: SUBM=+1