Auto-detect the OpenFlow Version and determine which OpenflowProvider to choose in Neutron.
Identifies the plugin that handles the Openflow Connection (OpenflowPlugin vs protocol_plugins.openflow).
Based on that detect if it is OF1.0 or OF1.3. ovsdb.of.version config variable takes the first priority though.

By having Auto-detect feature, it is simpler for the customer to deploy the OVSDB bundle without the need
to chase after config variables.

Change-Id: I8bdb45ea1d265afed5ffe79b0cb476e55c8c2196
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>