Updating ovsdb dependencies to the LATEST SNAPSHOT in Controller & Yangtools project
After the Hydrogen Release, we decided to keep the inter-project dependencies only with the RELEASE version of the bundles.
But, Since the weekly release schedule is not ratified and actual SNAPSHOT dependencies are growing, we dont have much choice
other than to depend on the SNAPSHOT versions of the bundles.

Once we have a streamlined process for weekly RELEASE in place, we can change our strategy to depend only on the RELEASE bundles.
Till then, we have to depend on SNAPSHOTs to make progress.

Change-Id: I68ee36c6ff418b20e3a7a255b6604246b2567aef
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>