Separate API and Implementation in OVSDB Neutron
It's regarded as a best practice in Java to separate API and
Implementation. In OSGi we can take this one step further and place API
and implementation in seperate bundles (a pattern used in the controller
project). This is what I have attempted in this commit.

- Create a new bundle 'neutron.impl'
- Leave interfaces in 'neutron' bundle
- Move implementations to 'neutron.impl' bundle
- Add new bundle to root pom.xml
- Add interfaces for AdminConfigManager, TenantNetworkManager and
  InternalNetworkManager
- Fix typo in a method in TenantNetworkManager s/tenn/ten/

This is a start, and work is not yet 100% complete, but as this is
disruptive I would like to start getting things merged sooner rather
than later.

Remaining ToDos:
- Consider moving Abstract Base Classes to API bundle
    - BaseHandler
    - ProviderNetworkManager

Change-Id: Ib238059bebbbd55c5cc2cdada88629c3a927bd9b
Signed-off-by: Dave Tucker <dave.j.tucker@hp.com>