Add support for vlan tenant isolation to ovsdb neutron using OpenFlow 1.3
This is intial push of support vlan tenant isolation with OpenFlow 1.3.

Sam has the 1.0 task and this is only for OF 1.3.

This diff includes the dynamic selection for eth interface. It chooses eth interface by parsing openvswitch table in ovsdb and use odl system config to get eth interface name if failed to find in ovs table. And the eth port is added to ovs bridge domain when the host has first vlan tenant network VM instance spawned.

Change-Id: I2b262bd80e4fd4cc0df57c4d427fea91c30a691e
Signed-off-by: Hsin-Yi Shen <hshen@redhat.com>