Update patch set 1
Patch Set 1:

Though I agree with the need to make it simple, the approach can be more inline with the existing distribution mechanism.

The long pending work on re-massaging OVSDB distribution is the actual work that need to be addressed.
The recommended approach is to download the pre-built distribution (from Integration branch maybe) and then include the needed plugins from our local repo. By doing that, we also get all the dependent plugins (such as OpenFlow1.3 and others in the future). Also, Custom building all of these dependent modules is unneccessary and is far simpler to just use the existing pre-built distribution.

Infact that is the distribution maven changes that i mentioned in the weekly call.

Patch-set: 1