Bump up the bundle versions for the bundles based on Semantic Versioning.
Since we are performing major rearchitecture of the base OVSDB library and it results in backward incompatible
API changes, the MAJOR version is bumped up. But the neutron and northbound bundles will face functionality impact
but not API impact & hence only the MINOR version is bumped up.

Care must be taken to make sure the master and topic/schema branche versions dont step on each oth versions.
It causes CI headaches.

Change-Id: Ib1b1810c54695e8e59917928e62cda43907a7af6
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>