Add Vagrantfile for OVSDB Test/Dev environment
This commit adds a Vagrantfile to the root of the repository. This
allows a developer to issue a single command to pull up a development
and test environment:

    vagrant up

This will create 4 x VMs with 2GB RAM each in Virtual Box. VMs can be
created individually if required.

1 x VM for OpenDaylight
1 x VM for Mininet
2 x VMs for Devstack (1 x Control/Compute, and 1 x Compute node)

See README.Vagrant for full instructions

Change-Id: Ieaaec6cde09ce9ecc7b31861f86695c662849f7d
Signed-off-by: Dave Tucker <dave.j.tucker@hp.com>