Add support for vlan tenant isolation to ovsdb using OpenFlow 1.0.
This is an initial push of the changes needed to support vlan tenant isolation.
Ignore the logger's with shague in them since they are for my debugging and will go away.

This is using OF 1.0. Hsin-Yi has the 1.3 task.

The br-eth for vlan and br-tun for tunnels has been combined to br-net. A single patch port
connects br-net to br-int.

Still outstanding is a method to get the eth mapping to provider networks. Neutron
has the following data when a network is created:

neutron net-create vlan-net --provider:network_type vlan --provider:segmentation_id 2001 --provider:physical_network physnet1

In offline configuration there is a bridge_mappings value to map the physnet to the provder network:

OVS_BRIDGE_MAPPINGS=physnet1:br-eth1,physnet3:br-eth3

We would need something similar. Possibly can use the other_config in the same way that
local_ip is used.

Signed-off-by: Sam Hague <shague@redhat.com>