Add ovsdb vlan support by using OF1.3
This change is to support vlan tenant isolation using OF1.3
This is using OF1.3 and Sam has the 1.0 change.

The changes of common neutron files are committed by Sam.
Most of the change here is for OF1.3 only.

This commit also has dependency on BUG 1020 "Multiple output actions is not allowed in the action list of apply-action instruction". The vlan and tunnel broadcast is broken due to BUG1020.
We don't want BUG 1020 to block the vlan feature commit and send this change for code review now.
Onve BUG 1020 is fixed, broadcast should work properly.

Change-Id: Ibd16404212cc1c4de21921477e4d1cf335a12b39
Signed-off-by: Hsin-Yi Shen <hshen@redhat.com>