Update patch set 6
Patch Set 6:

(2 comments)

Hi Sam,

In addition to the inline minor comments, there is an important item that needs to be addressed. When a functionality is added/modified, depending on the backward compatibility of the change, we need to update the major/minor/bug-fix version of the bundle. 

I propose to change the bundle version of ovsdb.neutron to 0.6.1-SNAPSHOT (please note that we are already using 0.6.0-SNAPSHOT in the topic/schema branch and hence I dont want to overstep the bundle versions & hence using an unique bundle version of 0.6.1-SNAPSHOT here.

Can you please update the ovsdb.neutron/pom.xml and the other pom.xml files that refer to the ovsdb.neutron ?

Also a kind request is to make sure a quick testing is done without any Vlan changes to make sure there are no collateral damages to the existing GRE/VXLAN implementation.

Thanks,
Madhu

Patch-set: 6
Label: Verified=0