Update patch set 8
Patch Set 8: Code-Review+2

Thanks for taking care of this.

1 comment to keep in mind for OF10 optimization.
The Tunnel delete happens only when the last network is deleted. This make sense for OF13 implementation, because we share the same tunnel port/interface across all the tenants. But for OF10 case, Tunnel is pinned to a tenant.
Hence it would make sense if we can handle the Network delete case to remove the Tunnel created for the OF10 case.
But that can be handled using another gerrit.

Patch-set: 8
Label: Code-Review=+2