Add support for vlan tenant isolation to ovsdb using OpenFlow 1.0.
This is an initial push of the changes needed to support vlan tenant isolation.
Ignore the logger's with shague in them since they are for my debugging and will go away.

This is using OF 1.0. Hsin-Yi has the 1.3 task.

The br-eth for vlan and br-tun for tunnels has been combined to br-net. A single patch port
connects br-net to br-int.

Functions to parse the physical interface name from the provider_mappings have been added.

Signed-off-by: Sam Hague <shague@redhat.com>
Change-Id: I369165f4f92665242e033454b49a5ee6a4ea0381