Update patch set 2
Patch Set 2:

We recently found a bug around this area & having an immutable object returned from the library layer towards the plugin and From plugin towards the Application is missing.

Being an involved fix, we decided to do it as part of the topic/schema which will avoid lot of merge conflicts when the branches collapses. 

Shall we avoid doing this change on the master, rather work on https://bugs.opendaylight.org/show_bug.cgi?id=905 in the topic/schema branch ?

Patch-set: 2