Update patch set 8
Patch Set 8:

It's convention to put the Vagrantfile at the root of the repository. It's also more convenient...

I also feel if it's in the root, people will use it. This is a shared dev/test environment that is a) quickly repeatable and b) something we can all maintain.

It also lowers the barrier of entry to new starters as they can "vagrant up" vs. 2-3 weeks of hitting the same old issues with DevStack.

Patch-set: 8