Moving ipfix column to set<UUID> in Bridge table.
Please note that these fixes to the fixed tables will all be deprecated once the topic/schema
branch is collapsed into the master. These fixes are encouraged into the master only to support
existing users. That is the only reason we are not spending time with UT/IT scripts for these.

Change-Id: I48603ac3b204d4fde676100e95edcd02761d7586
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>