Update patch set 5
Patch Set 5: Code-Review-1

(9 comments)

Thanks for getting this rebased so quickly Hsin-Yi.
This is looking good. A couple of minor nits in the comments.

With regard to comments, I'm a little bit of a comment purist and there are some comments here that describe what the code is doing vs. adding additional context. I'd rather remove these now, as it's these type of comments that don't get updated when the functionality changes, leaving us with a situation reminiscent of this

// This code adds 1 and 1 together
Integer result = 3 + 4;

-1'ing based on the fact that there are nits/comments to address but otherwise this is looking really good!

Patch-set: 5
Label: Code-Review=-1