Update patch set 2
Patch Set 2: Code-Review-1

The point of schema driven library is to avoid any such table definition in the classes. Having a free floating Table name in a test code is fine because we are anchoring on a schema to test out the schema functionality. Infact the plan is to isolate the test also based on different schemas but that is beyond the discussion of this change.
So, Please abondon this change.

Patch-set: 2
Label: Code-Review=-1