Update patch set 2
Patch Set 2: Code-Review-1

There is an existing TODO in this isReady() method and by changing the condition, it is not addressing the TODO. So I dont see an actual point in getting this in.
A better and complete fix would be to implement the timeout.

Without such a fix, it might result in the indefinite wait issue that we recently faced.

Patch-set: 2
Label: Code-Review=-1