Update patch set 3
Patch Set 3: Code-Review-1

We need the Library to be schema driven and not static table names. As you can see, we have OvsDBClientTestITTyped written especially for this case of making the life easier for applications. So, having a Typed class which makes life simpler for applications such as ovsdb.neutron will be on a layer that is above library. The Library itself should be kept clean. 

OvsDBClientTestITTyped was written with that in mind and to give an example of how one might implement a wrapper layer on top of the library to make it a particular schema such as OpenVSwitch or hardware_vtep, etc...

This is as per design.

Infact, I loath the presence of the string "openvswtich" as the schema name in this file :-) and I have an open comment on that. I also have a fix to remove any such static reference to these static typed schema objects inside of the library layer. The changes are in my stash for a long time and time for pushing that now.

Patch-set: 3
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1