BUG-1056 brocastflow is not proper modified after deleting interface
Current logic for flow mod is not correct and this fix intends to fix the flow mod logic.

When removing the output action due to port deleting, current logic just record number of output action type in the action list until the loop find the item we want to delete. So if the deleted item happens to be the first output action in the action list (while there are other output actions in the list), the number of output action type is 0 and flow is deleted. This is wrong behavior. The flow should be modified to remove the output action, not removed in this case.
 
This fix uses boolean flag instead. The flag is set to true if it find any output action other than the deleted one exists in the action list. So the flow only gets removed when no other output actions are found. This is the desired logic and behavior we want.

Signed-off-by: Hsin-Yi Shen <hshen@redhat.com>