Update patch set 3
Patch Set 3:

As per discussion on IRC this change will be abandoned. We keep the library as pure to RFC 7047 as possible, which means the current IT will be changed to not use the Open vSwitch schema, but to be totally schema independent.

Patch-set: 3
Reviewer: Gerrit User 1452 <1452@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0