Update patch set 2
Patch Set 2:

Got some issue to reply comments inline. So I reply the comments below.

Yes we need the index. The reason is the actions must set order properly after recent openflowplugin/openflowjava change. So here what we do is we get the index of deleted item. Actions after this deleted item also need to change the order accordingly.

For example, the original action list is: [order=0, output=1], [order=1, output=2], [order=2, output=5], If we delete action of "output=2", the new action list should be: [order=0, output=1], [order=1, output=5],

This is why we need the index and the loop below for order shifting.

Patch-set: 2