ovsdb has a dependency on artifacts from integration project, that shall be removed
- ovsdb has a dependency on artifact from integration project, while
integration project depends on ovsdb. This is a circular dependency
for the release process because we cannot release OVSDB without
releasing INTEGRATION and viceversa. To break the circular dependency
i'm proposing to build the ovsdb distribution starting from the
controller one adding the few deltas needed. Yes this cause some
duplication but i couldn't find a better way to break it. Suggestions
are more then welcome!

Change-Id: Ifed07caddce549a96a7f6eb92d6c977f2f938d5c
Signed-off-by: Giovanni Meo <gmeo@cisco.com>