Update patch set 1
Patch Set 1:

I'm of the opinion that this is more of a functional/integration test. It is dependent on Netty and does not map 1:1 with the methods in class under test.
The UT for JsonRPCDecoder should use a mock/stub for Netty.

Patch-set: 1
Reviewer: Gerrit User 1452 <1452@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0