Update patch set 1
Patch Set 1:

This test case is actually mocking those scenarios, partial records, full records etc using an actual socket itself since I was lazy in writing mocks for all such cases, plus this approach was self contained.  But that does not change the fact that we are actually unit testing the JsonDecoder. If there are not enough test coverage we can add the missing bits.  I would like think of unit test as something that tests a particualr unit of code, with dependencies that can be easily provided by the text fixture reliably and are fast. These could be mocks or stuff like this. Mocks would be ideal, but I though this way was easier, but I am open to having a good set of mocks in lieu of this.

Patch-set: 1
Reviewer: Gerrit User 952 <952@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0