Update patch set 1
Patch Set 1:

The "_" was to mention that the method is a temporary private method which will go away when we refactor the internals. Right now we have both old and new code in the same repo and this method is translating between those two bits and should not be in the code base at all. I agree, I can change it with a comment if you feel strongly about it, but its temporary.

Patch-set: 1
Reviewer: Gerrit User 952 <952@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0