Update patch set 2
Patch Set 2: Code-Review+2

This is not the final fix to address the multiValued column. As discussed with Ashwin, we will have a few follow up fixes that will make the library APIs more friendly to the apps and also the internals are fully clean of static typed classes.

Will merge the fix due to the upcoming fixes that are being tracked.

Patch-set: 2
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>