Update patch set 5
Patch Set 5:

UT = Does this method do what it's supposed to (in isolation)
IT = Do these classes work together as expected
Functional Test = Test the user story - Can I do a Transact operation?

In most cases IT are pretty close to the user story and therefore standalone FT can be ignored. In this case, we should have IT/FT for every single element in RFC7047 to make sure that we have covered our bases e.g How do I perform a "transact" operation.

These tests are just testing the logic of each of the methods in OvsDbClientImpl.

On re-reviewing this, it's clear to me that most of these UT's are not actually testing much other than verifying that the calls to other classes (OvsdbRPC for example) are made. My hope is as we start to add functionality within these methods (validation perhaps) that these tests become worthwhile. For now merging them is optional ;)

Patch-set: 5
Reviewer: Gerrit User 1452 <1452@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0