Update patch set 5
Patch Set 5:

Yes. Agreed. But do you think it would be more appropriate to mock a more realistic value ?
Say for example, if we have to do proper UT to cover all the
scenarios, we will be inputting data that should result in failure/error and the json returned should contain an error to indicate the failure.
Mocking empty data and testing for the response of empty data doesnt seem more readable. 

As you can see, my request is to not the code/logic. 
It is just mock the reply data  with a more appropriate value.

Patch-set: 5
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0