Removing the Open_vSwitch schema hardcoding in OvsDBClient.
Though it is an harmless constant string, it goes against the motivation of the library redesign to be
schema independent. Hence moving the schema string to a more appropriate place (to the applications),
in this case, to the Integration Test files.
Also, modified the initialize functionality to include a test to check for the list_dbs to contain Open_vSwitch
schema as all the tests in this file are based on Open_vSwitch schema tables.

Change-Id: I3210abf1fa18b5707a27b719f7d481b5de76d6be
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>