Bug 1147: Fix programming of NORMAL flow entry when using OF 1.3 Provider.
The OVSDB Plugin programs the br-int and br-ex bridges with the appropriate
flows as they are discovered by the plugin. The OpenFlow 1.0 provider
correctly installs the LLDP and NORMAL flows, but the OpenFlow 1.3 provider
does not install the NORMAL flow on the br-ex bridge (note: the NORMAL flows
are only installed on br-ex bridges, and not br-int bridges, when using
OpenFlow 1.3).

This patch fixes this by programming of the NORMAL flow on br-ex
bridges when using the OpenFlow 1.3 provider.`

Change-Id: I32e5179cc94ca7afb513a5a6fb6bc6c6dc5cf84d
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>