BUG-1188: Race-like behavior in programming of OpenStack instance flows causes flood+egress rules to not being programmed
In neutron’s SouthboundHandler::processRowUpdate() it is possible to receive interface and port updates in any order.
Both port and interface update code paths —provided they can find the neutronNetwork affected — end up calling:

   TenantNetworkManager::programTenantNetworkInternalVlan(Node node, String uuid, NeutronNetwork network);

And that updates port information as well as tenant network container.
In the case of port update, the provider was not being called back, which could render the provider out of date in
respect to the tenant’s ports. So, in short: we were missing a call to:

   this.handleInterfaceUpdate(node, intfUUID.toString(), into);

when processRowUpdate() was called for a port last. Because of that, provider would not add flow rules to handle
the port.

Change-Id: If3872985c7410cd48679129029cbf27c7d02eac9
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>