Initial hardware_vtep schema changes
These initial fixes wont consider IntegrationTests because vtep schema is not enabled in OpenVSwitch soft switches
and we need an vtep emulator to have this Integration Tested.

We need to invest extra effort in getting such a IT infra in place & hence the IT for hardware_vtep schema can be
considered once the hardware_vtep emulator setup is in place.

Change-Id: Ia60bac094a7ca3321be531ad3d16b9fd17ca18da
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>