Update patch set 4
Patch Set 4: Code-Review+2

The Docker IT failure for this commit is legitimate due to the IPFIX schema updates between version 1.10.0 and before. And the exception was appropriately raised from the code.

The test cases must be modified to make sure this expected result is honored rather than failing. Subsquent fix for ipfix test case will take care of it.

Patch-set: 4
Label: Code-Review=+2
Label: SUBM=+1