Plugin migration to use the new Schema independent Library.
Plugin layer is completely migrated to use the new Schema independent Library.
Since the existing plugin layer has a lot of Open_vSwitch assumption, this migration did not remove such an assumption
and break functionalities. Rather, a few design compromises were made in an effort to make this transition smoother.
Once these commits are merged, we will introduce more Schema independent APIs that will have no assumption on the
underlying Schemas.
Because of this reason, you will see a few places with access to OvsVswitchdSchemaConstants class, which indicates that
the Open_vSwitch schema assumption is made. Also, there are a few "special" case handling done on the ConfigurationService
to deal with existing API assumptions. These assumptions are marked as Special. Again, when the new APIs are introduced
these Assumptions can be easily removed.

The plugin layer changes helped in hardening the Library layer and hence resulted in some Library changes as well.

Enough effort is invested in both the PAX-EXAM based IT and regular Plugin IT code to ensure smooth migration.

Change-Id: Ie6365a23e3fa9b6ca4431d3eba7494330c8a07ca
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>