Update patch set 1
Patch Set 1:

(1 comment)

Please refer to http://docs.oracle.com/javase/tutorial/java/annotations/predefined.html and it will be clear that like Deprecated, Override is also a Compile time Retention policy which means it has no impact on the runtime behaviour of a method. It just indicates to the Compiler and IDEs to have more intelligent checks.

Just by the presence of the same Method signature, the child class will by default override the parent method of the same method signature and a super.<method-name> will make sure the parent method is called and that is already taken care in this method.

The presence or absence of @Override annotation will not impact it.

Patch-set: 1
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0