Update patch set 2
Patch Set 2: Code-Review-1

Hi Vishal,

We are in the process of Collapsing the topic/schema branch into the master and it is targetted to be done next week. Till that is complete, we are keeping master in the locked-down state. am just marking this as -1 to indicate the fellow committers not to merge it in master till the collapse is complete.

Also, we are thinking of removing the usage of ServiceHelper extensively and make the service references injected via DI. Hence this patch would have to undergo more changes to stick to the DI design choice.

I would suggest to wait for the collapse to be done and work on the DI based patch.

-Madhu

Patch-set: 2
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1