Bug 1327 : Fixing a few Connection handling issues
Since the library and plugin layers are seperated out into its own OSGi bundles and the initial low-level
OVSDB connection is managed by the library layer, there are cases in which the plugin layer might become
ACTIVE at a later point in time after the actual Passive OVSDB connection is established with the Library.
Also, the ConnectionManager and higher application layers can independently be restarted.

This fix brings in the robustness in the Connection management interactions between Library, Plugin, AD-SAL
and Connection Manager NSF.

In turn this makes the printCache command to work consistently which addresses the issue raised in Bug 1327

Change-Id: Ifc004f01e0306f7e71f908c063472a0b38bb572c
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>