Removed the executorservice thread that we added in async update processing
We moved channelRead in JsonRpc Layer to its own thread in order to avoid a possible channel-lock
issue when we try to program the table in the same thread.
But this breaks a lot of assumption in the existing code in Neutron/Library layers where an update
followed by access has to be delayed.

With the earler changes made to the Monitor request resulting in synchronous MonitorResponse, the
channel lock issue for the initial monitor processing is taken care of.

Any subsequent Updates are anyways thrown to an application queue which gets processed in its own
thread.

Hence, removing the thread from the low-layer and that enables receiving the monitor tableupdates/
updating the tableCache in the same thread.

Change-Id: I8a70b03fe87f8002bfc78a55b28d71c3e7152776
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>