Update patch set 1
Patch Set 1: Code-Review-1

(10 comments)

In addition to the inline comments, I also feel there are places some levels have to be changed.
For example : there are errors in netvirt code that are not errors and hence even in normal operations, it is filled with error logs (when we try to access a bridge like br-ex in compute nodes).

Can you please take care of them ?

Patch-set: 1
Label: Code-Review=-1