Update patch set 2
Patch Set 2:

(2 comments)

Verified locally and the tunnel port is not deleted.
This could be a regression from the neutron refactor as this behaviour is also present in master.
Will raise a bug to fix that but for now I think we are safe to merge this change

Patch-set: 2
Reviewer: Gerrit User 1452 <1452@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0